add group ignoration scheme to GlobalPotentials
group ignoration schemes are a method to detect groups that should be
ignored when calculate potential energy and forces (e.g. coarse-grained
electrostatics are not applied between particles that are in the same
chain). To generate neighbor lists, the potential provides it. Because
most of the methods are really simple, the schemes are passed as a
template parameter expecting compiler's optimization like inlining or
constant propagation.