use cast instead of math::constants
the reason of related commits is complicated. first, in C++, literal
`1.0` become `double`. if you want to write `float` literal, you need
to write `1.0f`. if you call a function like `T f(T, T)` with expression
`f(1.0 * x_in_float, y_in_float)`, template argument deduction end up
with a conflict between 'T is double' and 'T is float'. to avoid this,
we need to write literals that have correct types. but in a templatized
code, we cannot write a correct literal. we need type casting or
templatized constant values (or, if possible, explicitly specialize
function template. but some of the functions uses overloading).
and also it introduces (essentially needless) overhead related to type
casting like `cvtss2sd/cvtsd2ss` because `1.0 * x` becomes `double` even
if the type of `x` is `float`. finally I decided to use (like a C-style)
type casting for each constants. `static_cast` is too long to insert
everywhere we use constant. `real_type(1.0)` is cleaner than
`(real_type)1.0`. this is a result of compromise.