add neighbor_element for memory optimization
In C++, all the classes have at least 1 byte size even if the class has
no element. It's because C++ distinguishes object identity by an address.
There is a way to compress data size. By using Empty Base Optimization
(EBO), size of base class becomes zero. Based on this technique, the
memory usage in NeighborList can be optimized when pre-calculated
parameter is not needed. It uses less memory than the naive
implementation with std::pair<index, empty_class> (and also it can be
expected that it makes performance better because of cache locality).