dispatch ignore-by-chain dynamically
IgnoreChain provides a rule to ignore inter- or intra-chain interaction
by the Topology information. It would be callen only when it build an
ExclusionList that contains the pairs of indices in which no interaction
would be applied. So it never be a hotspot. To reduce the complexity,
IgnoreChain is removed from template argument and set at runtime.