fix: use const reference
Actually, it is a workaround for the auto-review comment and it does not
increase the efficiency. Since the function always copies the argument,
passing a const reference does not help to avoid copying. By moving the
argument, which is passerd by value, we don't need to have overloads for
`const&` and `&&` because both of the argument will be copied once and
moved into the member variable. However, const& can be used to receive
rvalue reference, we can use const&. To supress the comment, I added
const& here.