feat: add ContactInteraction
ContactInteraction is basically the same as BondLengthInteraction, but
it also has a list of active contacts. The list is maintained in the
same manner as VerletLists and broken contacts are safely skipped. Using
this with GoContactPotential and GaussianPotential would make the
simulation faster if a number of those contacts are broken throughout
the simulation.