perf: use normal NeighborList in CellList
ParallelNeighborList consumes quite a lot of memory resources and makes
both construction and access slower (especially when system has a small
number of particles). Because of this, after some benchmarking, I found
that normal NeighborList works good for most of the cases. Some part of
neighbor-list construction cannot be parallelized, but it becomes still
faster.