fix: omp::sort: copy to buf regardless of merging
To merge two sub-ranges that were sorted in parallel, it uses a buffer.
Although the range at the end of region does not need to merge, it
requries to be copied into buf from vec because the items in buf and vec
would be swapped after merge. It causes an error if it is executed with
non-power-of-two number of threads (e.g. 6 cores).