refactor: relax specialization condition
At first, this partial specialization conflicts with default
implementation of L-J and EXV, so I made the condition strict (exact
match with Debye-Huckel). But now the specialization conditions of the
default implementation of L-J and EXV are relaxed, I also relax the
condition here.
This change will reduce the implementation cost of global pair
potentials.