refactor: remove mostly meaningless reference
In most cases, matrix uses float or double. Those builtin small types
are smaller than or the same size as reference, that is internally a
pointer, so reference-passing-style does have an overhead. Although the
references should be removed after inlinize the operator by a compiler
while optimization process, I would make it clear.