style: suppress linter warnings around subprocess
by just splitting subprocess call into a shell script.
Of course it does not solve anything. It still calls a "subprocess", and
changing shell script has a "security issue" because one can execute any
command. But these scripts are just a utility script and will not run
automatically. I don't think it has a security issue because users can
read it and can decide to execute it or not.