refactor: introduce pair_parameter_type
distinguish parameter for each particle and parameter for a pair of
particles.

GlobalPotential has a list of parameters for each particle. Also, it
provides a way to calculate parameter for specific pair of particles
that probably interact with each other. This enables NeighborList to
cache values that are useful to calculate energy and force between
particles, e.g., sqrt(e_i * e_j) for Lennard-Jones, q_i * q_j / epsilon
for ElectroStatic. Generally, a parameter for a pair of particles can be
different from a parameter for each particle. So it is good to
distinguish them. This is the reason why `pair_parameter_type` is
introduced to global potentials.