perf: speedup initialization of FLP dihedral
Since it makes the minimum energy zero, it requires some initialization
to find the minimum point in the energy function and it takes relatively
long time. Although it does not effect to the overall efficiency because
it would be done only once, still it is painful to wait initialization.
This change caches the minimum energy value and make initialization
faster by look-up min energy that are already calculated.