refactor: use cerr directly in progress_bar
To remove complicated buffer size calculations. It may reduce runtime
efficiency. But progress bar can be disabled via input file and it
will automatically be turned off when the stderr is redirected. When we
run a job on a cluster using job scheduler, both stdout and stderr are
redirected. So it does not have large impact on the efficiency,
practically.