feat: add AFMFlexibleFittingInteraction
Note: this is not exactly the same code used in the paper, because we
used the older release version of Mjolnir and the code included some
experimental ideas. After submission, we polished the code a bit to gain
more efficiency. Thus there are some minor changes, including changes
made for meging.