feat: enable to load system status from msgpack
In the following way.
```toml
[[systems]]
file_name = "checkpoint.msg"
```
Still, I'm not fully convinced about this change. It introduces several
msgpack-related stuff to the read_system file. But developing a general
library seems not to be a nice option because it takes time and much
effort... If a nice idea came up with, we can replace it.