feat: change msgpack format of dynvars
Since default dynvar does not have boundary, it returns [-inf, inf]
as the lower and upper boundary. Msgpack can contain inf and nan
values of floating point, because it contains float/double as the
corresponding IEEE 754 format. All's right with the world.
    However, JSON cannot handle inf. People normally do not edit
msgpack itself, because it is a binary file, but might edit json
converted from msgpack. In that case, saving [-inf, inf] becomes
a problem. To avoid this, we omit lower/upper in the case of
default dynamic variables.