feat: add "fatal" level to log
to terminate immediately without throwing an exception
Some functions that normally do not throw any error are marked as
noexcept. But in some cases, the function cannot be calculated or are
kept unimplemented. In such a case, we need a way to halt the simulation
when such a function is called without throwing any error. LOG_FATAL
calls `std::terminate` after showing an error message.